Pictures added and game rezise
Added the image and gif for the shooter and the laser as well as rezised the whole game to make it easier to see.

I'd like to find a way to rezise the picture and gif of the shooter and laser cuz they're really small in order to fit into the grid.